Revert "Improve Basin redstone power handling by making better use of actual states, and not ddoing any NBT stuff."
This reverts commit 07d17e9389bf4f1bb27dde9337ab780cf21c310a.

Doesn't work because getActualState does not affect the state returned by world.getBlockState, so IS_OPEN is always the default value. This would either need to be converted to a real block state or simply reverted. I'm reverting it because I don't see the point in converting it to a real block state.